requirements.txt
asgiref==3.4.1
astroid==2.8.5
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-import-export==2.6.1
et-xmlfile==1.1.0
isort==5.10.1
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
platformdirs==2.4.0
pylint==2.11.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
typing_extensions==4.0.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
